Read HOST from env var
So we can configure it differently for production.

This uncovered the fact that we were actually performing HTTP
requests from the tests. Addin httpretty solves it raising an
UnmockedError. See: https://httpretty.readthedocs.io/en/latest/api.html?highlight=enable()#enable.

However, httpretty overlaps with responses package. Both aim to solve
the same thing, mocking http requests, and with similar APIs so we
better stick to httpretty to avoid an extra dependency. We can do that
later on though. Now we can go on with the implementation. I created
https://github.com/coopdevs/lazona_connector/issues/13 out of it.