Implement Koiki API client (#7)
* Implement create_delivery

* Handle HTTP response codes

* Log failed requests

We test by mocking the logger but in the future it might better to use
https://docs.python.org/3.4/library/unittest.html#unittest.TestCase.assertLogs.
Now it's just good enough as is.

* Fix linting errors

* Exclude Django files from coverage

* Treat address_2 as optional

* Read HOST from env var

So we can configure it differently for production.

This uncovered the fact that we were actually performing HTTP
requests from the tests. Addin httpretty solves it raising an
UnmockedError. See: https://httpretty.readthedocs.io/en/latest/api.html?highlight=enable()#enable.

However, httpretty overlaps with responses package. Both aim to solve
the same thing, mocking http requests, and with similar APIs so we
better stick to httpretty to avoid an extra dependency. We can do that
later on though. Now we can go on with the implementation. I created
https://github.com/coopdevs/lazona_connector/issues/13 out of it.

* DRY test setup moving it to setUp method

* Stub logging in test

This hides the `logging.error` call which outputs:

```
System check identified no issues (0 silenced).
...ERROR:root:Failed request. status=<MagicMock name='post().status_code' id='140169930956128'>, body=<MagicMock name='post().text' id='140169931312960'>
.....
----------------------------------------------------------------------
Ran 8 tests in 0.038s

OK
```

One may think this is a warning of some sort and it's not.