Configure DEBUG based on ENV var
This is copied from
https://developer.mozilla.org/en-US/docs/Learn/Server-side/Django/Deployment#getting_your_website_ready_to_publish.

> The value of the DEBUG will be True by default, but will only be False
if the value of the DJANGO_DEBUG environment variable is set to False.
Please note that environment variables are strings and not Python types.
We therefore need to compare strings. The only way to set the DEBUG
variable to False is to actually set it to the string False