Deal with koiki responses (#26)
* Follow Python method naming convention

* Consider failed responses any other than code 101

It feels safer to only consider a success what's explicitly returned as
a 101. As we won't explicitly catch those 102 response codes, leaving trace of
that as documentation is a good idea.

* Wrap Koiki responses into our own Delivery entity

This way we control how their API response's are translated so we can
adapt things in the future if needed. We no longer directly depend on
them.