Remove irrelevant tests
These may have been useful while TDDing what we have now, but as is,
they are not providing any value anymore. This file is only meant to
contain acceptance tests, so the less coupled it is to the
implementation the less we'll have to update it (we're doing that too
often now).