Handle vendors (#41)
* Serialize line_item metadata

* Remove irrelevant tests

These may have been useful while TDDing what we have now, but as is,
they are not providing any value anymore. This file is only meant to
contain acceptance tests, so the less coupled it is to the
implementation the less we'll have to update it (we're doing that too
often now).

* Build sender from line_item's metadata

* Build a delivery per vendor

Each delivery will then contain the aggregation of line_items that are
provided by each vendor.

* Remove unneeded attribute

* Allow line_items with more than vendor_id metadata

* Rename DeliverySerializer to OrderSerializer

We confused the system's input with its output. Woocommerce sends orders
regardless of the data we send to Koiki, which are indeed deliveries.

* Add doc

* Use POPOs instead of cumbersome dictionaries

This reduced coupling will ease future development.

* Move models into their own separate modules

This shows a clear distinction between the two sets of models and helps
clarify the design.

* Decouple Koiki's models from Woocommerce dicts

Recipient now relies on a duck type; any shipping or billing objects you
inject must conform to that public API.

* Raise exception if no vendor_id is found

Without that id there's nothing we can do, so better to fail loudly.

* Remove dead method

Now, instead of merging the two addresses, we pass them on separately as
well.

* Improve method documentation