Fetch vendor details (#45)
* Replace keys() + fetch with items()

* Fetch full vendor details when converting to dict

This acts like sort of a proxy. We don't place the HTTP request to get
all the details from WP's API until it's really needed.

* Fix indentation

* Log requests to Woocommerce

* Move responsibility to call get outside of model

* Extract WCFM Marketplace client to its namespace

And move logging to APIClient. This low level thing doesn't belong to
a model and neither to these tests.

* Parametrize WCFMMP API base

So we can easily switch from staging to production as well as testing.

* Explicitly pass Vendor id and name

* Remove now unused imports

* Remove unnecessary stubbed request

* Decouple Koiki's test config from test classes

* Move endpoint path from client to resource class

This would couple the API client with every endpoint we call and we
don't want that. Every resource object should care for its own endpoint.

* Refactor dict building with defaultdict

* Rename get() to fetch() to clarify its intent

* Replace hardcoded URL path with config var