Support partial ISO 3166-2 and fullname states
This abstraction will guard us from the complexity of both formats that
may come as input. It'll also make it possible to reuse this fallback
logic in other places of the app.

Note this also fixes the requests to Koiki that were failing because
Woocommerce provides `B` while Koiki expects `Barcelona`.