Read security-related settings vars from ENV (#67)
I expected Django to read these from ENV itself but it seems it does
not. For instance, while the ENV var is set
`settings.SECURE_SSL_REDIRECT` returns `False` from a shell.

I chose to only override the setting if the matching ENV var is set,
otherwise we default to Django's default value.